GEODE-4328: Make backup into a task run by a single thread
